Add values parameters for K8S Stardog probes
Add values parameters for K8S liveness and readiness probes. This can be
useful when running on different clouds where the time to provision may
vary. The previous settings are maintained as defaults. Also increase
the time that the post-install admin password setting waits for Stardog
to become available. Fixes #19.