Add startupProbe
It is currently necessary to change the
`livenessProbe.initialDelaySeconds` if stardog does not start fast
enough. This can be due to slow cloud storage or other reasons like a
index rebuild that takes longer than a normal startup. This means that
the database cannot start fast, if there

This adds [startupProbe]. The `startupProbe` is like the `livenessProbe`
but only until it succeeds once. Then it gets replaced by the
`livenessProbe`. This makes it possible to have no
`initialDelaySeconds` and low `periodSeconds` for the `livenessProbe` to
have fast failure detection while still allowing for long starup periods
with the `startupProbe` having a high `failureThreshold`.

[starupProbe]: https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#define-startup-probes